requirements.txt
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.2
dnspython==2.1.0
gunicorn==20.0.4